1. mybatis-plus升级到3.3.2 2. curator-recipes升级到5.0.0 3. zookeeper升级到3.6.1 4. fastjson升级到1.2.70 5. 并添加是否动态创建数据库的配置 6. 主题添加批量删除功能

+++ web/src/main/java/com/pegasus/kafka/controller/TopicController.java
@@ -167,8 +167,14 @@ public class TopicController {
 
     @PostMapping("del")
     @ResponseBody
-    public Result<?> del(@RequestParam(value = "topicName", required = true) String topicName) throws Exception {
-        kafkaTopicService.delete(topicName.trim());
+    public Result<?> del(@RequestParam(value = "topicNames", required = true) String topicNames) throws Exception {
+        String[] topicNameArray = topicNames.split(",");
+        for (String topicName : topicNameArray) {
+            if (StringUtils.isEmpty(topicName.trim())) {
+                continue;
+            }
+            kafkaTopicService.delete(topicName.trim());
+        }
         return Result.ok();
     }
 }
